backend/pyproject.toml
[tool.poetry]
name = "sd-get-pro"
version = "0.1.0"
description = ""
authors = ["desonglll <lindesong666@163.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
django-cors-headers = "^4.6.0"
djangorestframework-camel-case = "^1.4.2"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
